backend/requirements.txt
supabase==2.3.0
pydantic==2.5.2
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1